refactor: group ERC20 token code
