+++ src/evm/mod.rs
@@ -6,7 +6,6 @@ pub mod account_storage;
 pub mod engine_db;
 pub mod protocol;
 pub mod simulation;
-pub mod token;
 pub mod traces;
 pub mod tycho_models;
