+++ src/evm/protocol/vm/assets/ERC20.bin
