+++ src/evm/protocol/vm/mod.rs
@@ -1,9 +1,9 @@
 mod adapter_contract;
-pub mod constants;
-pub(crate) mod erc20_overwrite_factory;
+mod constants;
+mod erc20_token;
 mod models;
 pub mod state;
 pub mod state_builder;
 pub mod tycho_decoder;
-pub(crate) mod tycho_simulation_contract;
+mod tycho_simulation_contract;
 pub mod utils;
